backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0